+++ projects/demo/src/app/demo/pages/native-text-field-page/usage/native-text-field-usage.component.ts
@@ -1,8 +1,10 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { ChangeDetectionStrategy, Component, Input } from '@angular/core';
 
 @Component({
   selector: 'pa-demo-native-text-field-usage',
   templateUrl: './native-text-field-usage.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class NativeTextFieldUsageComponent {}
+export class NativeTextFieldUsageComponent {
+  @Input() native = false;
+}