+++ projects/pastanaga-angular/src/lib/controls/slider/slider.component.ts
@@ -1,15 +1,6 @@
-import {
-  ChangeDetectionStrategy,
-  Component,
-  EventEmitter,
-  Input,
-  OnChanges,
-  OnInit,
-  Output,
-  SimpleChanges,
-} from '@angular/core';
-
-let nextId = 0;
+import { AfterViewInit, ChangeDetectionStrategy, Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
+import { PaFormControlDirective } from '../form-field';
+import { coerceNumberProperty } from '@angular/cdk/coercion';
 
 @Component({
   selector: 'pa-slider',
@@ -17,45 +8,52 @@ let nextId = 0;
   styleUrls: ['./slider.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class SliderComponent implements OnInit, OnChanges {
-  @Input() id?: string;
-  @Input() name?: string;
-  @Input() min = 0;
-  @Input() max = 100;
-  @Input() step = 1;
-  @Input() help?: string;
-  @Input() isDisabled = false;
-  @Output() valueChange: EventEmitter<number> = new EventEmitter();
-  helpId = '';
-  rangeValue = 0;
-
-  ngOnInit() {
-    this.id = !this.id ? `field-range-${nextId++}` : `${this.id}-field-range`;
-    this.name = this.name || this.id;
-    this.helpId = `${this.id}-help`;
+export class SliderComponent extends PaFormControlDirective implements AfterViewInit, OnInit {
+  @Input()
+  set min(value: any) {
+    this._min = coerceNumberProperty(value);
   }
-
-  ngOnChanges(changes: SimpleChanges): void {
-    if (changes['min'] && typeof changes['min'].currentValue !== 'number') {
-      this.min = 0;
-    }
-
-    if (changes['max'] && typeof changes['max'].currentValue !== 'number') {
-      this.max = 100;
-    }
-
-    if (changes['step'] && typeof changes['step'].currentValue !== 'number') {
-      this.step = 5;
-    }
+  get min(): number {
+    return this._min;
   }
 
   @Input()
-  get value(): number {
-    return this.rangeValue;
+  set max(value: any) {
+    this._max = coerceNumberProperty(value);
+  }
+  get max(): number {
+    return this._max;
   }
 
-  set value(val) {
-    this.rangeValue = val;
-    this.valueChange.emit(this.rangeValue);
+  @Input()
+  set step(value: any) {
+    this._step = coerceNumberProperty(value);
+  }
+  get step(): number {
+    return this._step;
+  }
+
+  get outputMinWidth() {
+    return `${(this.max.toString().length + 1.5) * 8}px`;
+  }
+
+  @ViewChild('label', { read: ElementRef }) label?: ElementRef;
+  hasLabel = false;
+
+  private _min = 0;
+  private _max = 100;
+  private _step = 1;
+
+  ngAfterViewInit(): void {
+    this.hasLabel = !!this.label && this.label.nativeElement.textContent.length > 0;
+    this.cdr.markForCheck();
+  }
+
+  override ngOnInit() {
+    super.ngOnInit();
+    // Set value as the minimum one by default
+    if (!this.control.value) {
+      this.control.patchValue(this.min);
+    }
   }
 }