+++ projects/pastanaga-angular/src/lib/controls/slider/slider.module.ts
@@ -1,11 +1,13 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { FormsModule } from '@angular/forms';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { SliderComponent } from './slider.component';
+import { PaFormFieldModule } from '../form-field';
+import { PaFocusableModule } from '../../focusable';
 
 @NgModule({
   declarations: [SliderComponent],
-  imports: [CommonModule, FormsModule],
+  imports: [CommonModule, FormsModule, PaFormFieldModule, ReactiveFormsModule, PaFocusableModule],
   exports: [SliderComponent],
 })
 export class PaSliderModule {}